✨ Automatically detect churche when there is only one of its kind
When a place has only a single instance of the type we look for (a
church or a chapel) in clochers.org, we automatically pick that one.

It will help to detect "the only church of the city" when there are
small chapels as well.

This commit also adds the ability to fetch data from the existing record
in openstreetmap.

We incidentally set the building type when we can.